Fixing the Kotlin async function signatures
PollResult is an `i8` in Rust, but was incorrectly specified as a
`Short` in Kotlin.  Things seem to be working accidentally, but this
should be fixed.

+++ uniffi_bindgen/src/bindings/kotlin/templates/Helpers.kt
@@ -157,5 +157,5 @@ internal class UniFfiHandleMap<T: Any> {
 
 // FFI type for Rust future continuations
 internal interface UniFffiRustFutureContinuationCallbackType : com.sun.jna.Callback {
-    fun callback(continuationHandle: USize, pollResult: Short);
+    fun callback(continuationHandle: USize, pollResult: Byte);
 }
